Better embeds via script injection
Flourish provides better embeds, but they arent loaded in with iframes. Instead, the embeds are a div with a script tag inside, which gets executed when the DOM renders the div.
For whatever reason, React won't execute this script tag. To work around it, React refs are linked to these div embeds, and the script tag is "manually" inserted after the component has mounted
There's a special prefix to embed flourish prefixes. If written incorrectly, then the embed won't load.